// types/TranslatedAircraft.ts
import { Company, Airport, AircraftEngine, AircraftType } from ".";

export type TranslatedAircraft = {
    Id: string;
    Nickname: string;
    CurrentCompanyIdIfAny?: string;
    LastStatusChange: Date;
    CurrentStatusDurationInMinutes: number;
    AllowSell: boolean;
    AllowRent: boolean;
    AllowLease: boolean;
    SellPrice: number;
    RentHourPrice: number;
    RentFuelTotalGallons: number;
    RentCautionAmount: number;
    RentStartDate?: Date|null;
    RentLastDailyChargeDate?: Date|null;
    Identifier: string;
    Heading: number;
    Longitude: number;
    Latitude: number;
    FuelTotalGallons: number;
    FuelWeight: number;
    ZeroFuelWeight: number;
    AirframeHours: number;
    AirframeCondition: number;
    AirframeMaxCondition: number;
    LastAnnualCheckup?: Date|null;
    Last100hInspection?: Date|null;
    LastWeeklyOwnershipPayment?: Date|null;
    LastMagicTransportationDate?: Date|null;
    LastParkingFeePayment?: Date|null;
    IsControlledByAI: boolean;
    HoursBefore100HInspection: number;
    ConfigFirstSeats: number;
    ConfigBusSeats: number;
    ConfigEcoSeats: number;
    SeatsReservedForEmployees: number;
    ExtraWeightCapacity: number;
    TotalWeightCapacity: number;
    CurrentSeats: number;
    MustDoMaintenance: boolean;
    RentMaxDate?: Date|null;
    Altitude: number;
    FlightState: number;
    AircraftTypeId: string;
    WorldId: string;
    AircraftStatusId : string;
    RentalAirportId?: string;
    RentalCompanyId?: string;
    OnAirSyncedAt?: Date|null;
    Engines?: AircraftEngine[]|any;
    RentalAirport?: any;
    CurrentAirport?: any;
    RentalCompany?: any;
    CurrentCompany?: any;
    OwnerCompany?: any;
    AircraftType?: any;
    World?: any;
    AircraftStatus?: any;
}